--- app_server/src/main/java/knu/networksecuritylab/appserver/exception/book/BookDuplicateException.java
@@ -1,8 +1,0 @@
-package knu.networksecuritylab.appserver.exception.book;
-
-public class BookDuplicateException extends BookException {
-
-    public BookDuplicateException(BookErrorCode bookErrorCode) {
-        super(bookErrorCode);
-    }
-}
