--- app_server/src/main/java/knu/networksecuritylab/appserver/exception/book/BookNotFoundException.java
@@ -1,8 +1,0 @@
-package knu.networksecuritylab.appserver.exception.book;
-
-public class BookNotFoundException extends BookException {
-
-    public BookNotFoundException(BookErrorCode bookErrorCode) {
-        super(bookErrorCode);
-    }
-}
