+++ app_server/src/main/java/knu/networksecuritylab/appserver/exception/book/impl/BookNotFoundException.java
@@ -1,0 +1,11 @@
+package knu.networksecuritylab.appserver.exception.book.impl;
+
+import knu.networksecuritylab.appserver.exception.book.BookErrorCode;
+import knu.networksecuritylab.appserver.exception.book.BookException;
+
+public class BookNotFoundException extends BookException {
+
+    public BookNotFoundException(BookErrorCode bookErrorCode) {
+        super(bookErrorCode);
+    }
+}
