--- app_server/src/main/java/knu/networksecuritylab/appserver/exception/file/FileStorageException.java
@@ -1,8 +1,0 @@
-package knu.networksecuritylab.appserver.exception.file;
-
-public class FileStorageException extends FileException {
-
-    public FileStorageException(FileErrorCode fileErrorCode) {
-        super(fileErrorCode);
-    }
-}
