--- app_server/src/main/java/knu/networksecuritylab/appserver/exception/file/ImageNotFoundException.java
@@ -1,8 +1,0 @@
-package knu.networksecuritylab.appserver.exception.file;
-
-public class ImageNotFoundException extends FileException {
-
-    public ImageNotFoundException(FileErrorCode fileErrorCode) {
-        super(fileErrorCode);
-    }
-}
