--- app_server/src/main/java/knu/networksecuritylab/appserver/exception/file/ImageNotReadable.java
@@ -1,8 +1,0 @@
-package knu.networksecuritylab.appserver.exception.file;
-
-public class ImageNotReadable extends FileException {
-
-    public ImageNotReadable(FileErrorCode fileErrorCode) {
-        super(fileErrorCode);
-    }
-}
