--- app_server/src/main/java/knu/networksecuritylab/appserver/exception/user/InvalidAuthenticationException.java
@@ -1,8 +1,0 @@
-package knu.networksecuritylab.appserver.exception.user;
-
-public class InvalidAuthenticationException extends UserException {
-
-    public InvalidAuthenticationException(UserErrorCode userErrorCode) {
-        super(userErrorCode);
-    }
-}
