--- app_server/src/main/java/knu/networksecuritylab/appserver/exception/user/InvalidUsernameOrPassword.java
@@ -1,8 +1,0 @@
-package knu.networksecuritylab.appserver.exception.user;
-
-public class InvalidUsernameOrPassword extends UserException {
-
-    public InvalidUsernameOrPassword(UserErrorCode userErrorCode) {
-        super(userErrorCode);
-    }
-}
