--- app_server/src/main/java/knu/networksecuritylab/appserver/exception/user/UserNotFoundException.java
@@ -1,8 +1,0 @@
-package knu.networksecuritylab.appserver.exception.user;
-
-public class UserNotFoundException extends UserException {
-
-    public UserNotFoundException(UserErrorCode userErrorCode) {
-        super(userErrorCode);
-    }
-}
