--- app_server/src/main/java/knu/networksecuritylab/appserver/exception/user/UsernameDuplicateException.java
@@ -1,8 +1,0 @@
-package knu.networksecuritylab.appserver.exception.user;
-
-public class UsernameDuplicateException extends UserException {
-
-    public UsernameDuplicateException(UserErrorCode userErrorCode) {
-        super(userErrorCode);
-    }
-}
