+++ app_server/src/test/java/knu/networksecuritylab/appserver/controller/UserControllerTest.java
@@ -4,9 +4,9 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import knu.networksecuritylab.appserver.config.jwt.JwtAuthenticationFilter;
 import knu.networksecuritylab.appserver.controller.user.dto.SignInRequestDto;
 import knu.networksecuritylab.appserver.controller.user.dto.SignUpRequestDto;
-import knu.networksecuritylab.appserver.exception.user.UsernameDuplicateException;
+import knu.networksecuritylab.appserver.exception.user.impl.UsernameDuplicateException;
 import knu.networksecuritylab.appserver.exception.user.UserErrorCode;
-import knu.networksecuritylab.appserver.service.user.BasicUserService;
+import knu.networksecuritylab.appserver.service.user.impl.BasicUserService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
